feat(ui): extract shared toolbar and filter-panel styled components to @core
Move ToolButton, ToolButtonSlot, ToolDivider, ActiveToolButton to
@core/components/toolbar and SectionBox, SectionTitle, SectionDescription,
FilterGrid to @core/components/filter-panel so they are reusable project-wide.
Student list TableHeader now imports from these shared locations.
